server: validate user has permissions

+++ server/src/http/routes/delete-exercise.ts
@@ -1,11 +1,12 @@
 import { db } from '@/db'
 import { exercises } from '@/db/schema'
-import { eq } from 'drizzle-orm'
+import { and, eq } from 'drizzle-orm'
 import type { FastifyPluginAsyncZod } from 'fastify-type-provider-zod'
 import z from 'zod'
+import { auth } from '../middlewares/auth'
 
 export const deleteExercise: FastifyPluginAsyncZod = async app => {
-  app.delete(
+  app.register(auth).delete(
     '/exercises/:id',
     {
       schema: {
@@ -14,8 +15,18 @@ export const deleteExercise: FastifyPluginAsyncZod = async app => {
         }),
       },
     },
-    async request => {
+    async (request, reply) => {
       const { id } = request.params
+      const userId = await request.getCurrentUserId()
+
+      const [exercise] = await db
+        .select()
+        .from(exercises)
+        .where(and(eq(exercises.id, id), eq(exercises.userId, userId)))
+
+      if (!exercise) {
+        return reply.code(401).send({ message: 'Unauthorized' })
+      }
 
       await db.delete(exercises).where(eq(exercises.id, id))
     }
