+++ server/src/http/routes/delete-workout.ts
@@ -1,11 +1,12 @@
 import { db } from '@/db'
 import { workouts } from '@/db/schema'
-import { eq } from 'drizzle-orm'
+import { and, eq } from 'drizzle-orm'
 import type { FastifyPluginAsyncZod } from 'fastify-type-provider-zod'
 import z from 'zod'
+import { auth } from '../middlewares/auth'
 
 export const deleteWorkout: FastifyPluginAsyncZod = async app => {
-  app.delete(
+  app.register(auth).delete(
     '/workouts/:id',
     {
       schema: {
@@ -14,8 +15,18 @@ export const deleteWorkout: FastifyPluginAsyncZod = async app => {
         }),
       },
     },
-    async request => {
+    async (request, reply) => {
       const { id } = request.params
+      const userId = await request.getCurrentUserId()
+
+      const [workout] = await db
+        .select()
+        .from(workouts)
+        .where(and(eq(workouts.id, id), eq(workouts.userId, userId)))
+
+      if (!workout) {
+        return reply.code(401).send({ message: 'Unauthorized' })
+      }
 
       await db.delete(workouts).where(eq(workouts.id, id))
     }
